Move perflog testing to a different file
